+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/init/AzureFunctionsInitializerTestHolder.java
@@ -4,13 +4,7 @@
 package com.microsoft.applicationinsights.agent.internal.init;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
 
-import com.microsoft.applicationinsights.agent.internal.exporter.AgentLogExporter;
-import com.microsoft.applicationinsights.agent.internal.telemetry.TelemetryClient;
 import org.junit.jupiter.api.Test;
 
 class AzureFunctionsInitializerTestHolder {
@@ -18,21 +12,14 @@ class AzureFunctionsInitializerTestHolder {
   /*
    * Lazily Set Connection String For Linux Consumption Plan:
    *
-   *    Term      LazySetOptIn   ConnectionString      EnableAgent        LazySet
-   *    Preview   FALSE          VALID                 TRUE               Enabled
-   *                             VALID                 FALSE              Disabled
-   *                             VALID                 NULL               Disabled
-   *                             NULL                  TRUE/FALSE/NULL    Disabled
-   *    GA        TRUE           VALID                 TRUE               Enabled
-   *                             VALID                 FALSE              Disabled
-   *                             VALID                 NULL               Enabled
-   *                             NULL                  TRUE/FALSE/NULL    Disabled
+   *    Term      LazySetOptIn   EnableAgent        LazySet
+   *    Preview   FALSE          TRUE               Enabled
+   *                             FALSE              Disabled
+   *                             NULL               Disabled
+   *    GA        TRUE           TRUE               Enabled
+   *                             FALSE              Disabled
+   *                             NULL               Enabled
    */
-  private static final String CONNECTION_STRING =
-      "InstrumentationKey=00000000-0000-0000-0000-0FEEDDADBEEF";
-
-  private static final String INSTRUMENTATION_KEY = "00000000-0000-0000-0000-0FEEDDADBEEF";
-  private static final String WEBSITE_SITE_NAME = "fake_site_name";
 
   @Test
   // "LazySetOptIn is FALSE, ConnectionString is valid and EnableAgent is TRUE"
@@ -55,67 +42,15 @@ class AzureFunctionsInitializerTestHolder {
   @Test
   // "LazySetOptIn is FALSE, ConnectionString is NULL, InstrumentationKey is NULL, and EnableAgent
   // is TRUE"
-  void disableLazySetWithLazySetOptInOffConnectionStringNullInstrumentationKeyNull() {
+  void disableLazySetWithLazySetOptInOff() {
     assertThat(AzureFunctionsInitializer.isAgentEnabled("true", false)).isTrue();
-
-    // given
-    TelemetryClient telemetryClient = mock(TelemetryClient.class);
-    AgentLogExporter agentLogExporter = mock(AgentLogExporter.class);
-    AppIdSupplier appIdSupplier = mock(AppIdSupplier.class);
-    AzureFunctionsInitializer lazyConfigurationAccessor =
-        new AzureFunctionsInitializer(telemetryClient, agentLogExporter, appIdSupplier);
-
-    // when
-    lazyConfigurationAccessor.setConnectionString(null, null);
-
-    // then
-    verify(telemetryClient, never()).updateConnectionStrings(any(), any(), any());
-  }
-
-  @Test
-  // "LazySetOptIn is FALSE, ConnectionString is valid, InstrumentationKey is NULL, and EnableAgent
-  // is TRUE"
-  void disableLazySetWithLazySetOptInOffConnectionStringNotNullInstrumentationKeyNull() {
-    assertThat(AzureFunctionsInitializer.isAgentEnabled("true", false)).isTrue();
-
-    // given
-    TelemetryClient telemetryClient = mock(TelemetryClient.class);
-    AgentLogExporter agentLogExporter = mock(AgentLogExporter.class);
-    AppIdSupplier appIdSupplier = mock(AppIdSupplier.class);
-    AzureFunctionsInitializer lazyConfigurationAccessor =
-        new AzureFunctionsInitializer(telemetryClient, agentLogExporter, appIdSupplier);
-
-    // when
-    lazyConfigurationAccessor.setConnectionString(CONNECTION_STRING, null);
-
-    // then
-    verify(telemetryClient).updateConnectionStrings(CONNECTION_STRING, null, null);
-
-    // when
-    lazyConfigurationAccessor.setWebsiteSiteName(WEBSITE_SITE_NAME);
-
-    // then
-    verify(telemetryClient).updateRoleName(WEBSITE_SITE_NAME);
   }
 
   @Test
   // "LazySetOptIn is FALSE, ConnectionString is NULL, InstrumentationKey is valid, and EnableAgent
   // is TRUE")
-  void enableLazySetWithLazySetOptInOffConnectionStringNullInstrumentationKeyNotNull() {
+  void enableLazySetWithLazySetOptInOff() {
     assertThat(AzureFunctionsInitializer.isAgentEnabled("true", false)).isTrue();
-
-    // given
-    TelemetryClient telemetryClient = mock(TelemetryClient.class);
-    AgentLogExporter agentLogExporter = mock(AgentLogExporter.class);
-    AppIdSupplier appIdSupplier = mock(AppIdSupplier.class);
-    AzureFunctionsInitializer lazyConfigurationAccessor =
-        new AzureFunctionsInitializer(telemetryClient, agentLogExporter, appIdSupplier);
-
-    // when
-    lazyConfigurationAccessor.setConnectionString(null, INSTRUMENTATION_KEY);
-
-    // then
-    verify(telemetryClient).updateConnectionStrings(CONNECTION_STRING, null, null);
   }
 
   @Test
@@ -139,60 +74,14 @@ class AzureFunctionsInitializerTestHolder {
   @Test
   // "LazySetOptIn is TRUE, ConnectionString is NULL, InstrumentationKey is NULL, and EnableAgent is
   // TRUE"
-  void disableLazySetWithLazySetOptInOnConnectionStringNullAndInstrumentationKeyNull() {
+  void disableLazySetWithLazySetOptInOn() {
     assertThat(AzureFunctionsInitializer.isAgentEnabled("true", true)).isTrue();
-
-    // given
-    TelemetryClient telemetryClient = mock(TelemetryClient.class);
-    AgentLogExporter agentLogExporter = mock(AgentLogExporter.class);
-    AppIdSupplier appIdSupplier = mock(AppIdSupplier.class);
-    AzureFunctionsInitializer lazyConfigurationAccessor =
-        new AzureFunctionsInitializer(telemetryClient, agentLogExporter, appIdSupplier);
-
-    // when
-    lazyConfigurationAccessor.setConnectionString(null, null);
-
-    // then
-    verify(telemetryClient, never()).updateConnectionStrings(any(), any(), any());
   }
 
   @Test
   // "LazySetOptIn is TRUE, ConnectionString is valid, InstrumentationKey is NULL, and EnableAgent
   // is TRUE"
-  void enableLazySetWithLazySetOptInOnConnectionStringNotNullInstrumentationKeyNull() {
+  void enableLazySetWithLazySetOptInOn() {
     assertThat(AzureFunctionsInitializer.isAgentEnabled("true", false)).isTrue();
-
-    // given
-    TelemetryClient telemetryClient = mock(TelemetryClient.class);
-    AgentLogExporter agentLogExporter = mock(AgentLogExporter.class);
-    AppIdSupplier appIdSupplier = mock(AppIdSupplier.class);
-    AzureFunctionsInitializer lazyConfigurationAccessor =
-        new AzureFunctionsInitializer(telemetryClient, agentLogExporter, appIdSupplier);
-
-    // when
-    lazyConfigurationAccessor.setConnectionString(CONNECTION_STRING, null);
-
-    // then
-    verify(telemetryClient).updateConnectionStrings(CONNECTION_STRING, null, null);
-  }
-
-  @Test
-  // "LazySetOptIn is TRUE, ConnectionString is NULL, InstrumentationKey is valid, and EnableAgent
-  // is TRUE"
-  void enableLazySetWithLazySetOptInOnConnectionStringNullInstrumentationKeyNotNull() {
-    assertThat(AzureFunctionsInitializer.isAgentEnabled("true", false)).isTrue();
-
-    // given
-    TelemetryClient telemetryClient = mock(TelemetryClient.class);
-    AgentLogExporter agentLogExporter = mock(AgentLogExporter.class);
-    AppIdSupplier appIdSupplier = mock(AppIdSupplier.class);
-    AzureFunctionsInitializer lazyConfigurationAccessor =
-        new AzureFunctionsInitializer(telemetryClient, agentLogExporter, appIdSupplier);
-
-    // when
-    lazyConfigurationAccessor.setConnectionString(null, INSTRUMENTATION_KEY);
-
-    // then
-    verify(telemetryClient).updateConnectionStrings(CONNECTION_STRING, null, null);
   }
 }